Sets, Cleaning, Logic problems
Started working on a method to solve for numbers that can only be
figured out by nifty logic (As far as I know. I wish there were another
way...)
Basic solving is now described in terms of sets which makes things
simpler. Appropriate methods have been added.
A "list to hashable" function was added which may be useful for
"logicising".
Added a puzzle. Various cleaning things. I think the "was this the same
as the previous 'solve'" may actually be working now.